requirements.txt
numpy==1.11.1
Pillow==3.3.0